// src/app/api/models/activosfijos.ts
/* tslint:disable */
/* eslint-disable */
export interface Activosfijos {
  clasificacion?: string;
  depreciacionanual?: number;
  depreciacionmensual?: number;
  descripcion: string;
  empleadoasignado?: string;
  empleadosId?: string;
  estado?: boolean;
  fechaadquisicion?: string;
  id?: string;
  metodo?: string;
  ubicacion?: string;
  valoractual?: number;
  valorinicial: number;
  vidautil?: number;
}