// src/app/api/models/depreciacion.ts
/* tslint:disable */
/* eslint-disable */
export interface Depreciacion {
  activofijo: string;
  activosfijosId?: string;
  comentario?: string;
  depreciacionanual?: number;
  depreciacionmensual?: number;
  depreciaciontotal?: number;
  descripcion?: string;
  fechaadquisicion?: string;
  id?: string;
  metodo?: string;
  valoractual?: number;
  valorinicial: number;
  vidautil: number;
}